requirements.txt
cryptography==42.0.8
dnspython==2.6.1
loguru==0.7.2
jwcrypto==1.5.6
python-dotenv==1.0.1
requests==2.31.0
tldextract==5.1.2